support numpy arrays (only) and move to meshes.tsx
